Modified Scene dataclass to allow memory image representation
The Scene dataclass used to rely on the path attributes to perform operations on the tensors; it seems unnecessary given that the tensors aren't None. This would allow one to set tensors directly without having to provide paths to files (say someone using ROS topics and reads images from a topic message).